refactor(codegen): extrai analysis/address_taken.rs de func.rs
Move collect_address_taken_fns + scan_stmt + scan_expr (225 linhas)
para crates/rts-codegen/src/codegen/lower/analysis/address_taken.rs.
Conservador detector de fns cujo endereco eh tomado (sem TCO).

func.rs: 9385 -> 9161 linhas. Suite verde.
