+++ crates/rts-codegen/src/codegen/lower/analysis/mod.rs
@@ -2,4 +2,5 @@
 //!
 //! Submodulos folha (sem deps cruzadas entre si) extraidos de `func.rs`.
 
+pub mod address_taken;
 pub mod types;
